+++ Makefile
@@ -8,9 +8,9 @@ CFLAGS = -Wall -std=c99 -lm
 # Diretórios de código fonte e destino
 SRC_DIR = src
 PRIM_DIR = src/primitives
-COMP_DIR = src/compressor
+COMP_DIR = src/compression
 FMAN_DIR = src/fileManager
-TEST_DIR = src/test
+IMG_DIR = src/image
 OBJ_DIR = obj
 
 # Nome do binário final
@@ -22,7 +22,7 @@ SRC_FILES = $(wildcard $(SRC_DIR)/main.c) \
             $(wildcard $(PRIM_DIR)/*.c) \
             $(wildcard $(COMP_DIR)/*.c) \
             $(wildcard $(FMAN_DIR)/*.c) \
-			$(wildcard $(TEST_DIR)/*.c)
+            $(wildcard $(IMG_DIR)/*.c) \
 
 # Convertendo os arquivos .c em arquivos .o correspondentes
 # Cada diretório de origem será convertido para um caminho correto para o diretório de objetos
