+++ src/compression/dct.c
@@ -1,63 +1,63 @@
 #include "dct.h"
 
-    //============================
-    // TABELAS DA DCT
-    //============================
+//============================
+// TABELAS DA DCT
+//============================
 
-    // Matriz dos cossenos pré-calculada
-    const double C[8][8] = {
-        {0.354, 0.354, 0.354, 0.354, 0.354, 0.354, 0.354, 0.354},
-        {0.490, 0.416, 0.278, 0.098, -0.098, -0.278, -0.416, -0.490},
-        {0.462, 0.191, -0.191, -0.462, -0.462, -0.191, 0.191, 0.462},
-        {0.416, -0.098, -0.490, -0.278, 0.278, 0.490, 0.098, -0.416},
-        {0.354, -0.354, -0.354, 0.354, 0.354, -0.354, -0.354, 0.354},
-        {0.278, -0.490, 0.098, 0.416, -0.416, -0.098, 0.490, -0.278},
-        {0.191, -0.462, 0.462, -0.191,-0.191, 0.462,   -0.462, 0.191},
-        {0.098, -0.278, 0.416, -0.490, 0.490, -0.416,   0.278, -0.098}
-    };
+// Matriz dos cossenos pré-calculada
+const double C[8][8] = {
+    {0.354, 0.354, 0.354, 0.354, 0.354, 0.354, 0.354, 0.354},
+    {0.490, 0.416, 0.278, 0.098, -0.098, -0.278, -0.416, -0.490},
+    {0.462, 0.191, -0.191, -0.462, -0.462, -0.191, 0.191, 0.462},
+    {0.416, -0.098, -0.490, -0.278, 0.278, 0.490, 0.098, -0.416},
+    {0.354, -0.354, -0.354, 0.354, 0.354, -0.354, -0.354, 0.354},
+    {0.278, -0.490, 0.098, 0.416, -0.416, -0.098, 0.490, -0.278},
+    {0.191, -0.462, 0.462, -0.191,-0.191, 0.462,   -0.462, 0.191},
+    {0.098, -0.278, 0.416, -0.490, 0.490, -0.416,   0.278, -0.098}
+};
 
-    // Matriz dos cossenos transposta pré-calculada
-    const double CT[8][8] = {
-        {0.354, 0.490, 0.462, 0.416, 0.354, 0.278, 0.191, 0.098 }, 
-        {0.354, 0.416, 0.191, -0.098, -0.354, -0.490, -0.462, -0.278}, 
-        {0.354, 0.278, -0.191, -0.490, -0.354, 0.098, 0.462, 0.416}, 
-        {0.354, 0.098, -0.462, -0.278, 0.354, 0.416, -0.191, -0.490}, 
-        {0.354, -0.098, -0.462, 0.278, 0.354, -0.416, -0.191, 0.490}, 
-        {0.354, -0.278, -0.191, 0.490, -0.354, -0.098, 0.462, -0.416}, 
-        {0.354, -0.416, 0.191, 0.098, -0.354, 0.490, -0.462, 0.278}, 
-        {0.354, -0.490, 0.462, -0.416, 0.354, -0.278, 0.191, -0.098} 
-    };
-
-
-    //============================
-    // TABELAS DA QUANTIZAÇÃO
-    //============================
-
-    // Matriz de quantização para o componente da luminância
-    const int QY[8][8] = {
-        {16, 11, 10, 16, 24, 40, 51, 61},
-        {12, 12, 14, 19, 26, 58, 60, 55},
-        {14, 13, 16, 24, 40, 57, 69, 56},
-        {14, 17, 22, 29, 51, 87, 80, 62},
-        {18, 22, 37, 56, 68, 109, 103, 77},
-        {24, 35, 55, 64, 81, 104, 113, 92},
-        {79, 64, 78, 87, 103, 121, 120, 101},
-        {72, 92, 95, 98, 112, 100, 103, 99}
-
-    };
+// Matriz dos cossenos transposta pré-calculada
+const double CT[8][8] = {
+    {0.354, 0.490, 0.462, 0.416, 0.354, 0.278, 0.191, 0.098 }, 
+    {0.354, 0.416, 0.191, -0.098, -0.354, -0.490, -0.462, -0.278}, 
+    {0.354, 0.278, -0.191, -0.490, -0.354, 0.098, 0.462, 0.416}, 
+    {0.354, 0.098, -0.462, -0.278, 0.354, 0.416, -0.191, -0.490}, 
+    {0.354, -0.098, -0.462, 0.278, 0.354, -0.416, -0.191, 0.490}, 
+    {0.354, -0.278, -0.191, 0.490, -0.354, -0.098, 0.462, -0.416}, 
+    {0.354, -0.416, 0.191, 0.098, -0.354, 0.490, -0.462, 0.278}, 
+    {0.354, -0.490, 0.462, -0.416, 0.354, -0.278, 0.191, -0.098} 
+};
 
 
-    // Matriz de quantização para os componentes de croôminancia
-    const int QC[8][8] = {
-        {17, 18, 24, 47, 99, 99, 99, 99},
-        {18, 21, 26, 66, 99, 99, 99, 99},
-        {24, 26, 56, 99, 99, 99, 99, 99},
-        {47, 66, 99, 99, 99, 99, 99, 99},
-        {99, 99, 99, 99, 99, 99, 99, 99},
-        {99, 99, 99, 99, 99, 99, 99, 99},
-        {99, 99, 99, 99, 99, 99, 99, 99},
-        {99, 99, 99, 99, 99, 99, 99, 99}
-    };
+//============================
+// TABELAS DA QUANTIZAÇÃO
+//============================
+
+// Matriz de quantização para o componente da luminância
+const int QY[8][8] = {
+    {16, 11, 10, 16, 24, 40, 51, 61},
+    {12, 12, 14, 19, 26, 58, 60, 55},
+    {14, 13, 16, 24, 40, 57, 69, 56},
+    {14, 17, 22, 29, 51, 87, 80, 62},
+    {18, 22, 37, 56, 68, 109, 103, 77},
+    {24, 35, 55, 64, 81, 104, 113, 92},
+    {79, 64, 78, 87, 103, 121, 120, 101},
+    {72, 92, 95, 98, 112, 100, 103, 99}
+
+};
+
+
+// Matriz de quantização para os componentes de croôminancia
+const int QC[8][8] = {
+    {17, 18, 24, 47, 99, 99, 99, 99},
+    {18, 21, 26, 66, 99, 99, 99, 99},
+    {24, 26, 56, 99, 99, 99, 99, 99},
+    {47, 66, 99, 99, 99, 99, 99, 99},
+    {99, 99, 99, 99, 99, 99, 99, 99},
+    {99, 99, 99, 99, 99, 99, 99, 99},
+    {99, 99, 99, 99, 99, 99, 99, 99},
+    {99, 99, 99, 99, 99, 99, 99, 99}
+};
 
 
 //============================
