+++ src/fileManager/fileManager.h
@@ -4,28 +4,22 @@
     #include <stdlib.h>
     #include <stdbool.h>
 
-    #include "bfHeader.h"
-    #include "bfInfoHeader.h"
     #include "cmpHeader.h"
-    #include "bmp.h"
-
+    
     #include "../primitives/bitBuffer.h"
     #include "../primitives/vector.h"
     #include "../primitives/matrix.h"
-    #include "../primitives/pixel.h"
     #include "../primitives/util.h"
-
-    #include "../compressor/sampler.h"
-    #include "../compressor/dct.h"
+   
+    #include "../image/bfHeader.h"
+    #include "../image/bfInfoHeader.h"
+    #include "../image/pixel.h"
+    #include "../image/bmp.h"
     
     // Funções de I/O para arquivos
     BMP *loadBmpImage(char *bmpName);
     bool writeBmpImage(char *bmpName, BMP *newImage);
 
+    bitBuffer *loadCmpFile(char *binName, BIHEADER *iHeader, BFHEADER *fHeader, CMPHEADER *cHeader);
     bool writeCmpFile(char *binName, BIHEADER iHeader, BFHEADER fHeader, CMPHEADER cHeader, BITBUFFER *cmpData);
-
-    // Funções de compressão / descompressão
-    //bool compress(BMP *bmp);
-    //bool decompress();
-
 #endif
