+++ src/image/bfInfoHeader.c
