+++ src/image/bfInfoHeader.h
