+++ src/image/bmp.c
